Move DetektJvmSpec tests that rely on TestKit to functionalTest
