+++ detekt-gradle-plugin/src/test/kotlin/io/gitlab/arturbosch/detekt/DetektJvmSpec.kt
@@ -64,81 +64,5 @@ class DetektJvmSpec {
                 assertThat(argumentString).contains("--classpath")
             }
         }
-
-        @Nested
-        inner class `report location set on extension & task` {
-            val gradleRunner = DslGradleRunner(
-                projectLayout = ProjectLayout(numberOfSourceFilesInRootPerSourceDir = 1),
-                buildFileName = "build.gradle.kts",
-                mainBuildFileContent = """
-                    plugins {
-                        kotlin("jvm")
-                        id("io.gitlab.arturbosch.detekt")
-                    }
-
-                    repositories {
-                        mavenCentral()
-                        mavenLocal()
-                    }
-
-                    detekt {
-                        reports {
-                            txt.destination = file("output-path.txt")
-                        }
-                    }
-
-                    tasks.withType<io.gitlab.arturbosch.detekt.Detekt>().configureEach {
-                        reports {
-                            txt.destination = file("output-path2.txt")
-                        }
-                    }
-                """.trimIndent(),
-                dryRun = false
-            ).also {
-                it.setupProject()
-            }
-
-            @Test
-            fun `logs a warning`() {
-                gradleRunner.runTasksAndCheckResult(":detektMain") { buildResult ->
-                    assertThat(buildResult.output).contains("TXT report location set on detekt {} extension will be ignored for detektMain task.")
-                }
-            }
-        }
-
-        @Nested
-        inner class `report location set on task only` {
-            val gradleRunner = DslGradleRunner(
-                projectLayout = ProjectLayout(numberOfSourceFilesInRootPerSourceDir = 1),
-                buildFileName = "build.gradle.kts",
-                mainBuildFileContent = """
-                    plugins {
-                        kotlin("jvm")
-                        id("io.gitlab.arturbosch.detekt")
-                    }
-
-                    repositories {
-                        mavenCentral()
-                        mavenLocal()
-                    }
-
-                    tasks.withType<io.gitlab.arturbosch.detekt.Detekt>().configureEach {
-                        reports {
-                            txt.destination = file("output-path2.txt")
-                        }
-                    }
-                """.trimIndent(),
-                dryRun = false
-            ).also {
-                it.setupProject()
-            }
-
-            @Test
-            fun `logs a warning`() {
-                gradleRunner.runTasksAndCheckResult(":detektMain") { buildResult ->
-                    assertThat(buildResult.output).doesNotContain("report location set on detekt {} extension will be ignored")
-                }
-            }
-        }
     }
 }
